# services/scenario-manager/docs/swagger.yaml
basePath: /
definitions:
  entities.ComputeConfig:
    properties:
      name:
        type: string
      number_of_compute_nodes:
        type: integer
      number_of_port_per_vm:
        type: integer
      number_of_vm_per_vpc:
        type: integer
      scheduler:
        type: string
      vm_deploy_type:
        type: string
      vpc_info:
        items:
          $ref: '#/definitions/entities.VPCInfo'
        type: array
    type: object
...
  entities.Service:
    properties:
      cmd:
        type: string
      name:
        type: string
      parameters:
        items:
          type: string
        type: array
      return_code:
        items:
          type: integer
        type: array
      return_string:
        items:
          type: string
        type: array
      url:
        type: string
      when_to_run:
        type: string
      where_to_run:
        type: string
    type: object
...
  entities.ServiceConfig:
    properties:
      name:
        type: string
      services:
        items:
          $ref: '#/definitions/entities.Service'
        type: array
    type: object
  entities.SubnetInfo:
    properties:
      number_of_vms:
        type: integer
      subnet_cidr:
        type: string
      subnet_gateway:
        type: string
    type: object
...
  entities.VPCInfo:
    properties:
      number_of_subnets:
        type: integer
      project_id:
        type: string
      subnet_info:
        items:
          $ref: '#/definitions/entities.SubnetInfo'
        type: array
      tenant_id:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Restful APIs for Merak Cloud Emulator
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Merak Rest API
  version: "2.0"
paths:
...
  /api/compute-config:
    get:
      consumes:
      - application/json
      description: Get all compute-config
      responses:
        "200":
          description: array of compute-config with success message
          schema:
            items:
              $ref: '#/definitions/entities.ComputeConfig'
            type: array
        "404":
          description: null compute-config data with error message
      summary: Get all compute-config from database
      tags:
      - compute-config
    post:
      consumes:
      - application/json
      description: Create a compute-config
      parameters:
      - description: ComputeConfig
        in: body
        name: compute_config
        required: true
        schema:
          $ref: '#/definitions/entities.ComputeConfig'
      responses:
        "200":
          description: Compute data with success message
          schema:
            $ref: '#/definitions/entities.ComputeConfig'
        "500":
          description: Compute null with failure message
      summary: Insert a compute-config to database
      tags:
      - compute-config
  /api/compute-config/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a compute-config
      parameters:
      - description: ComputeConfId
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: compute-config data with success message
          schema:
            $ref: '#/definitions/entities.ComputeConfig'
        "404":
          description: compute-config data with null and error message
      summary: Delete a compute-config from database
      tags:
      - compute-config
    get:
      consumes:
      - application/json
      description: Get a compute-config
      parameters:
      - description: ComputeConfId
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: compute-config data with success message
          schema:
            $ref: '#/definitions/entities.ComputeConfig'
        "404":
          description: compute-config data with null and error message
      summary: Get a compute-config from database
      tags:
      - compute-config
    put:
      consumes:
      - application/json
      description: Update a compute-config
      parameters:
      - description: ComputeConfId
        in: path
        name: id
        required: true
        type: string
      - description: ComputeConfig
        in: body
        name: compute_config
        required: true
        schema:
          type: string
      responses:
        "200":
          description: compute_config data with success message
          schema:
            $ref: '#/definitions/entities.ComputeConfig'
        "500":
          description: compute_config null with failure message
      summary: Update a compute-config to database
      tags:
      - compute-config
...
  /api/service-config:
    get:
      consumes:
      - application/json
      description: Get all service-config
      responses:
        "200":
          description: array of service-config with success message
          schema:
            items:
              $ref: '#/definitions/entities.ServiceConfig'
            type: array
        "404":
          description: null service-config data with error message
      summary: Get all service-config from database
      tags:
      - service-config
    post:
      consumes:
      - application/json
      description: Create a service-config
      parameters:
      - description: ServiceConfig
        in: body
        name: service_config
        required: true
        schema:
          $ref: '#/definitions/entities.ServiceConfig'
      responses:
        "200":
          description: service-config data with success message
          schema:
            $ref: '#/definitions/entities.ServiceConfig'
        "500":
          description: service-config null with failure message
      summary: Insert a service-config to database
      tags:
      - service-config
  /api/service-config/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a service-config
      parameters:
      - description: NetworkId
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: service-config data with success message
          schema:
            $ref: '#/definitions/entities.ServiceConfig'
        "404":
          description: service-config data with null and error message
      summary: Delete a service-config from database
      tags:
      - service-config
    get:
      consumes:
      - application/json
      description: Get a service-config
      parameters:
      - description: NetworkId
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: service-config data with success message
          schema:
            $ref: '#/definitions/entities.ServiceConfig'
        "404":
          description: service-config data with null and error message
      summary: Get a service-config from database
      tags:
      - service-config
    put:
      consumes:
      - application/json
      description: Update a service-config
      parameters:
      - description: NetworkId
        in: path
        name: id
        required: true
        type: string
      - description: ServiceConfig
        in: body
        name: Service_config
        required: true
        schema:
          type: string
      responses:
        "200":
          description: service-config data with success message
          schema:
            $ref: '#/definitions/entities.ServiceConfig'
        "500":
          description: service-config null with failure message
      summary: Update a service-config to database
      tags:
      - service-config
...
schemes:
- http
swagger: "2.0"